remote_a2a/facts_agent/__main__.py
from google.adk.a2a.utils.agent_to_a2a import to_a2a
from .agent import root_agent

# Let to_a2a create the entire app
a2a_app = to_a2a(root_agent, port=8001)

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(a2a_app, host="0.0.0.0", port=8001)