requirements.txt
google-adk==1.18.0
a2a-sdk==0.3.11